+++ cosmosid/api/reports.py
@@ -28,27 +28,16 @@ class RunReportResponseStatus:
 class Reports(object):
     _resource_path = '/api/metagenid/v2/files/report/tsv'
 
-    def __init__(self, base_url=None, api_key=None, file_id=None, run_id=None):
+    def __init__(self, base_url=None, api_key=None, file_id=None):
         self.base_url = base_url
         self.logger = LOGEGR
         self.header = {'X-Api-Key': api_key}
         self.report_type = None
         self.file_id = file_id
-        self.run_id = run_id
         self.run_o = Runs(base_url=self.base_url,
                           api_key=self.header['X-Api-Key'])
         self.session = requests_retry_session(self.header)
         self.session.headers.update()
-
-
-    def __is_runid_in_file(self):
-        """Get given run meta and check is the run in sample."""
-        single_run = self.run_o.get_single_run(self.run_id)
-        if single_run:
-            if single_run['status']:
-                if single_run['file']['id'] == self.file_id:
-                    return True
-        return False
 
     def await_report_task(self, task_id, timeout=5*60):
         task_url = f"{self.base_url}/api/metagenid/v2/files/report/{task_id}"
@@ -81,10 +70,6 @@ class Reports(object):
                                         self.__class__._resource_path)
             if not self.file_id:
                 raise ValidationError('File ID is required')
-            if self.run_id and not self.__is_runid_in_file():
-                    raise NotFound('Given Run id {} is not in '
-                                   'given File {}'.format(self.run_id,
-                                                          self.file_id))
 
             # TODO: propagate supported type and tax_level
             params = {"files": [self.file_id]}